+++ system/src/utils/DishonoredUtils.mjs
@@ -25,7 +25,7 @@ export default class DishonoredUtils {
 			const systemVersion = game.system.version;
 
 			if (systemVersion !== savedVersion) {
-				Hotbar.toggleDocumentSheet(
+				foundry.applications.ui.Hotbar.toggleDocumentSheet(
 					CONFIG.DISHONORED.JOURNAL_UUIDS.releaseNotes
 				);