+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [#54] Reduce actor/item sheet brightness to help with eye strain
 
 ## Chores
+- [#52] Merged French translation updates from Crowdin
 - [#53] Momentum/Chaos tracker too large and overlaps Foundry v13 UI elements
 
 	* Redesigned and reimplemented as an ApplicationV2 based application that takes up less screen real estate