+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#55] Accessing the global "ActorSheet" which is now namespaced under foundry.appv1.sheets.ActorSheet
 - [#56] Accessing the global "ItemSheet" which is now namespaced under foundry.appv1.sheets.ItemSheet
 - [#57] Accessing the global "Actors" which is now namespaced under foundry.documents.collections.Actors
+- [#58] Accessing the global "Items" which is now namespaced under foundry.documents.collections.Items
 
 ---