+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#57] Accessing the global "Actors" which is now namespaced under foundry.documents.collections.Actors
 - [#58] Accessing the global "Items" which is now namespaced under foundry.documents.collections.Items
 - [#59] Update node modules used by build process
+- [#60] Accessing the global "loadTemplates" which is now namespaced under foundry.applications.handlebars.loadTemplates
 
 ---