+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [#58] Accessing the global "Items" which is now namespaced under foundry.documents.collections.Items
 - [#59] Update node modules used by build process
 - [#60] Accessing the global "loadTemplates" which is now namespaced under foundry.applications.handlebars.loadTemplates
+- [#61] Accessing the global "Token" which is now namespaced under foundry.canvas.placeables.Token
 
 ---